Move initial_state property to EvolvedOpAnsatz
